Explicit type conversions

+++ BASIC/src/basic_interpreter.cpp
@@ -1349,7 +1349,7 @@ Interpreter::readInput()
 		return false;
 
 	const uint8_t availableSize = PROGSTRINGSIZE - 1 - _inputPosition;
-	a = min(a, availableSize);
+	a = min(uint8_t(a), availableSize);
 
 	const size_t read = _input.readBytes(_inputBuffer + _inputPosition, a);
 	assert(read <= availableSize);
